Some small refactoring in OverviewPage to reduce LOC
